% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{cwNA}
\alias{cwNA}
\title{Check dataframe for NA}
\usage{
cwNA(x)
}
\arguments{
\item{x}{dataframe}
}
\value{
logical. If TRUE, then the dataframe contains an NA.
}
\description{
Returns TRUE if a data frame is NA
}
